Animation refactoring
Logical refactoring to fit more with usual game design. Change the step function into an engine method as it essentially calls update and render. Keep the animation function in global scope because it stores window functions. And then create a self-invoking function which handles the animation and stepping. Also rename the GameEngine object to simply just Engine