Refactor menus to inherit from Menu object
Put the types of menus in the same style as other objects in the game: var Object = function() {.... The key thing was to ensure that Game/Menu objects are not instantiated when the options for the menu are created. This means that the newScreen is just a regular js object and not a specialised game one.