Put the event listener on the document instead of canvas
This simple change (well return, my previous poor commit doesn't specify why the event handlers were put on the canvas element in the first place) drastically improves the frame rate of the game. It was observed that performance increased when the canvas wasn't the focus e.g. when the cursor was on the address bar. So first I removed the event listeners from the canvas element to the document element and then I removed the tab index. Performance is back at its best